+++ include/wx/qt/choice.h
@@ -19,7 +19,7 @@ public:
             const wxPoint& pos = wxDefaultPosition,
             const wxSize& size = wxDefaultSize,
             int n = 0, const wxString choices[] = (const wxString *) NULL,
-            long style = 0,
+            unsigned int style = 0,
             const wxValidator& validator = wxDefaultValidator,
             const wxString& name = wxASCII_STR(wxChoiceNameStr) );
 
@@ -27,7 +27,7 @@ public:
             const wxPoint& pos,
             const wxSize& size,
             const wxArrayString& choices,
-            long style = 0,
+            unsigned int style = 0,
             const wxValidator& validator = wxDefaultValidator,
             const wxString& name = wxASCII_STR(wxChoiceNameStr) );
 
@@ -35,7 +35,7 @@ public:
             const wxPoint& pos = wxDefaultPosition,
             const wxSize& size = wxDefaultSize,
             int n = 0, const wxString choices[] = NULL,
-            long style = 0,
+            unsigned int style = 0,
             const wxValidator& validator = wxDefaultValidator,
             const wxString& name = wxASCII_STR(wxChoiceNameStr) );
 
@@ -43,7 +43,7 @@ public:
             const wxPoint& pos,
             const wxSize& size,
             const wxArrayString& choices,
-            long style = 0,
+            unsigned int style = 0,
             const wxValidator& validator = wxDefaultValidator,
             const wxString& name = wxASCII_STR(wxChoiceNameStr) );
