+++ include/wx/qt/dataview.h
@@ -73,7 +73,7 @@ public:
 
     wxDataViewCtrl( wxWindow *parent, wxWindowID id,
            const wxPoint& pos = wxDefaultPosition,
-           const wxSize& size = wxDefaultSize, long style = 0,
+           const wxSize& size = wxDefaultSize, unsigned int style = 0,
            const wxValidator& validator = wxDefaultValidator );
 
     virtual ~wxDataViewCtrl();
@@ -81,7 +81,7 @@ public:
 
     bool Create(wxWindow *parent, wxWindowID id,
            const wxPoint& pos = wxDefaultPosition,
-           const wxSize& size = wxDefaultSize, long style = 0,
+           const wxSize& size = wxDefaultSize, unsigned int style = 0,
            const wxValidator& validator = wxDefaultValidator );
 
     virtual bool AssociateModel( wxDataViewModel *model );
