+++ include/wx/qt/minifram.h
@@ -24,7 +24,7 @@ public:
               const wxString& title,
               const wxPoint& pos = wxDefaultPosition,
               const wxSize& size = wxDefaultSize,
-              long style = wxCAPTION | wxCLIP_CHILDREN | wxRESIZE_BORDER,
+              unsigned int style = wxCAPTION | wxCLIP_CHILDREN | wxRESIZE_BORDER,
               const wxString& name = wxASCII_STR(wxFrameNameStr))
   {
       return wxFrame::Create(parent, id, title, pos, size,
@@ -37,7 +37,7 @@ public:
               const wxString& title,
               const wxPoint& pos = wxDefaultPosition,
               const wxSize& size = wxDefaultSize,
-              long style = wxCAPTION | wxCLIP_CHILDREN | wxRESIZE_BORDER,
+              unsigned int style = wxCAPTION | wxCLIP_CHILDREN | wxRESIZE_BORDER,
               const wxString& name = wxASCII_STR(wxFrameNameStr))
   {
       Create(parent, id, title, pos, size, style, name);
