+++ include/wx/qt/msgdlg.h
@@ -17,7 +17,7 @@ class WXDLLIMPEXP_CORE wxMessageDialog : public wxMessageDialogBase
 public:
     wxMessageDialog(wxWindow *parent, const wxString& message,
                     const wxString& caption = wxASCII_STR(wxMessageBoxCaptionStr),
-                    long style = wxOK|wxCENTRE,
+                    unsigned int style = wxOK|wxCENTRE,
                     const wxPoint& pos = wxDefaultPosition);
     virtual ~wxMessageDialog();
