+++ include/wx/qt/slider.h
@@ -19,7 +19,7 @@ public:
              int value, int minValue, int maxValue,
              const wxPoint& pos = wxDefaultPosition,
              const wxSize& size = wxDefaultSize,
-             long style = wxSL_HORIZONTAL,
+             unsigned int style = wxSL_HORIZONTAL,
              const wxValidator& validator = wxDefaultValidator,
              const wxString& name = wxASCII_STR(wxSliderNameStr));
 
@@ -28,7 +28,7 @@ public:
                 int value, int minValue, int maxValue,
                 const wxPoint& pos = wxDefaultPosition,
                 const wxSize& size = wxDefaultSize,
-                long style = wxSL_HORIZONTAL,
+                unsigned int style = wxSL_HORIZONTAL,
                 const wxValidator& validator = wxDefaultValidator,
                 const wxString& name = wxASCII_STR(wxSliderNameStr));
